Reorganized tests for further expansion
